move sync service into a separate module
