+++ crates/p2p/src/lib.rs
@@ -5,15 +5,16 @@ mod gossip_service;
 mod peer_list;
 mod server;
 mod server_data_handler;
+mod sync;
 #[cfg(test)]
 mod tests;
 mod types;
+
 pub use gossip_client::GossipClient;
 pub use gossip_service::fast_forward_vdf_steps_from_block;
-pub use gossip_service::sync_chain;
 pub use gossip_service::P2PService;
 pub use gossip_service::ServiceHandleWithShutdownSignal;
-pub use gossip_service::SyncState;
 pub use peer_list::{PeerListFacade, PeerListFacadeError, PeerListServiceFacade};
 pub use peer_list::{PeerListService, PeerListServiceError};
+pub use sync::{sync_chain, SyncState};
 pub use types::{GossipError, GossipResult};
